script+tg_bot/requirements.txt
aiofiles==24.1.0
aiogram==2.25
aiohappyeyeballs==2.4.3
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
frozenlist==1.5.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
propcache==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
pyTelegramBotAPI==4.23.0
python-certifi-win32==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.7
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools-scm==8.1.0
six==1.16.0
sniffio==1.3.1
tomli==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.17.1


